+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## Bugfixes
 
-* None
+* Fixed typo within class name "TraitDefinition"
 
 ## Features
 
 * Introduced assertions which allow for unwrapped error messages to not expose system internals
+* Added structure and method based override for enforcement processing configuration based on annotations
 
 # Version 1.1.0